Update index.html
trying option 2 (suggested by chat gpt) to raising the text vertically. Changed "justiify-content: center;" to "justify-content: flex-start;" which he explained would bring the text (numbers sequence) all the way to the very tippy top of the page, which looks bad, so I had to also add "padding-top: 100px;" which clearly adds invisible "padding" (think of it as a margin) from the top of the page, which would therefore bring the text down.